+++ src/KappaDuck.Quack/UI/Progress/IProgressOperation.cs
@@ -1,7 +1,7 @@
 // Copyright (c) KappaDuck.
 // Licensed under the MIT license.
 
-namespace KappaDuck.Quack.Progress;
+namespace KappaDuck.Quack.UI.Progress;
 
 /// <summary>
 /// Represents a progress operation that can be reported or cancelled.
