+++ src/KappaDuck.Quack/UI/Progress/Reporters/ProgressReporter.cs
@@ -1,7 +1,7 @@
 // Copyright (c) KappaDuck.
 // Licensed under the MIT license.
 
-namespace KappaDuck.Quack.Progress;
+namespace KappaDuck.Quack.UI.Progress.Reporters;
 
 /// <summary>
 /// Synchronous determinate progress reporter handed to <see cref="ProgressBar.Start(Action{ProgressReporter}, int)"/>.
