+++ tests/Unit.Tests/UI/Progress/Reporters/AsyncIndeterminateProgressReporterTests.cs
@@ -1,10 +1,10 @@
 // Copyright (c) KappaDuck.
 // Licensed under the MIT license.
 
-using KappaDuck.Quack.Progress;
+using KappaDuck.Quack.UI.Progress.Reporters;
 using System.Diagnostics.CodeAnalysis;
 
-namespace Unit.Tests.Progress;
+namespace Unit.Tests.UI.Progress.Reporters;
 
 internal sealed class AsyncIndeterminateProgressReporterTests : IDisposable
 {
